refactor: move to RuleTypes
